Guerrero_William_prueba_rpa/requirements.txt
beautifulsoup4==4.12.3
certifi==2026.4.22
charset-normalizer==3.4.7
et_xmlfile==2.0.0
idna==3.13
numpy==2.4.4
openpyxl==3.1.2
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
requests==2.31.0
six==1.17.0
soupsieve==2.8.3
tzdata==2026.2
urllib3==2.6.3
